HEALTHCARE_system/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-migrate==4.0.5
flask-cors==4.0.0
werkzeug==3.0.1
twilio==8.10.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0